test(rag): cover fail-closed boundary lines in CI-selected suite
