+++ tests/test_philosophy_pipeline.py
@@ -997,97 +997,3 @@ class TestRunPipeline:
         assert result.filtered_chunks == []
         assert [stage.stage_name for stage in result.stage_results] == ["rule_validation"]
         assert result.post_stage1_enrichment_completed is False
-
-
-# ===========================================================================
-# Fail-closed boundary lines (diff-coverage sentinels)
-# ===========================================================================
-
-
-class TestFailClosedBoundaryLines:
-    """Focused coverage for the fail-closed and advisory boundary lines."""
-
-    def test_post_init_rejects_completed_enrichment_without_filtered_chunks(self) -> None:
-        """Keyword construction of the impossible empty-and-complete state raises."""
-        with pytest.raises(
-            ValueError,
-            match="post-Stage-1 enrichment cannot complete without filtered chunks",
-        ):
-            PipelineResult(
-                filtered_chunks=[],
-                stage_results=[],
-                warnings=[],
-                total_latency_ms=0.0,
-                post_stage1_enrichment_completed=True,
-            )
-
-    def test_post_init_accepts_completed_enrichment_with_survivors(self) -> None:
-        """Completed enrichment with survivors is the only valid completed state."""
-        result = PipelineResult(
-            filtered_chunks=[_chunk()],
-            stage_results=[],
-            warnings=[],
-            total_latency_ms=0.0,
-            post_stage1_enrichment_completed=True,
-        )
-
-        assert result.post_stage1_enrichment_completed is True
-        assert len(result.filtered_chunks) == 1
-
-    def test_stage1_exception_returns_sanitized_fail_closed_result(
-        self,
-        caplog: pytest.LogCaptureFixture,
-    ) -> None:
-        """Stage-1 boundary exception rejects everything without leaking internals."""
-        with (
-            caplog.at_level(logging.WARNING, logger="core.rag.philosophy_pipeline"),
-            patch(
-                "core.rag.philosophy_pipeline._stage1_rule_validation",
-                side_effect=RuntimeError("sentinel-stage1-exception"),
-            ),
-        ):
-            result = run_pipeline([_chunk()], "sentinel-stage1-query")
-
-        assert result.filtered_chunks == []
-        assert result.stage_results == []
-        assert result.warnings == ["validation_error: internal failure, no chunks accepted"]
-        assert result.post_stage1_enrichment_completed is False
-        assert result.total_latency_ms >= 0
-        assert "Stage-1 RAG validation failed; rejecting all chunks" in caplog.text
-        assert "sentinel-stage1-exception" not in caplog.text
-        assert "sentinel-stage1-query" not in caplog.text
-        assert caplog.records
-        assert all(record.exc_info is None for record in caplog.records)
-
-    def test_alignment_mismatch_flagged_through_full_pipeline(self) -> None:
-        """High-score short Stage-1 survivors are flagged by Stage 3 end-to-end."""
-        chunks = [
-            _chunk("c1", "Short but valid.", 0.95, "a.md"),
-            _chunk(
-                "c2",
-                "A normal chunk with sufficiently long wellness content.",
-                0.6,
-                "b.md",
-            ),
-        ]
-        result = run_pipeline(chunks, "wellness query")
-
-        assert result.post_stage1_enrichment_completed is True
-        stage3 = result.stage_results[2]
-        assert stage3.stage_name == "source_alignment"
-        assert stage3.metadata == {"flagged_count": 1}
-        assert result.warnings == ["alignment_mismatch"]
-
-    def test_numeric_contradiction_counted_through_full_pipeline(self) -> None:
-        """Contradictory anchored ranges are counted by Stage 4 end-to-end."""
-        chunks = [
-            _chunk("c1", "Normal BP range is 90-120 for adults.", 0.9, "a.md"),
-            _chunk("c2", "Normal BP range is 140-180 for adults.", 0.8, "b.md"),
-        ]
-        result = run_pipeline(chunks, "What BP range is normal?")
-
-        assert result.post_stage1_enrichment_completed is True
-        stage4 = result.stage_results[3]
-        assert stage4.stage_name == "logical_consistency"
-        assert stage4.metadata == {"unique_sources": 2, "contradiction_count": 1}
-        assert result.warnings == ["numeric_contradiction"]
